refactor: move AddCategory usecase to commands folder
